+++ src/view/empty-pathes-list.js
@@ -1,15 +1,17 @@
 import { createElement } from '../utils';
-import { createFormTemplate } from '../form-template.js';
 
-export default class NewForm {
-  constructor(tasks) {
-    this._tasks = tasks;
-    this._resetButtonName = 'Cancel',
+const createEmptyListTemplate = () => (
+  `<p class="trip-events__msg">
+    Click New Event to create your first point
+  </p>`);
+
+export default class EmptyList {
+  constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createFormTemplate(this._tasks, this._resetButtonName);
+    return createEmptyListTemplate();
   }
 
   getElement() {
